# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "crm-gateway",
        "version": "fix2"
    },
    "paths": {
        "/v1/image/edit": {
            "post": {
                "tags": [
                    "image"
                ],
                "summary": "Apply an allow-listed edit operation",
                "description": "The single entry point for every generative edit.\n\nYou send an **op name**, not a prompt \u2014 the gateway owns the wording (see\nops.py). Fixed-function ops (upscale, restore, remove-background) answer\ninline; diffusion ops return `202 {job_id}` and you poll `/v1/jobs/{id}`.",
                "operationId": "edit_v1_image_edit_post",
                "parameters": [
                    {
                        "name": "wait",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "minimum": 0.0,
                            "description": "Seconds to block waiting for a diffusion result before returning a job id. 0 (default) returns 202 immediately. Capped by SYNC_WAIT_MAX.",
                            "default": 0,
                            "title": "Wait"
                        },
                        "description": "Seconds to block waiting for a diffusion result before returning a job id. 0 (default) returns 202 immediately. Capped by SYNC_WAIT_MAX."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageOut"
                                }
                            }
                        }
                    },
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobAccepted"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/image/upscale": {
            "post": {
                "tags": [
                    "image"
                ],
                "summary": "Real-ESRGAN upscale (2x or 4x)",
                "operationId": "upscale_v1_image_upscale_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpscaleIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/image/remove-bg": {
            "post": {
                "tags": [
                    "image"
                ],
                "summary": "Background removal (U^2-Net)",
                "operationId": "remove_bg_v1_image_remove_bg_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveBgIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/ocr": {
            "post": {
                "tags": [
                    "ocr"
                ],
                "summary": "PP-OCRv6 text detection + recognition",
                "operationId": "ocr_v1_ocr_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OcrIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OcrOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/audio/transcribe": {
            "post": {
                "tags": [
                    "audio"
                ],
                "summary": "Whisper transcription",
                "operationId": "transcribe_v1_audio_transcribe_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TranscribeIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranscribeOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/audio/denoise": {
            "post": {
                "tags": [
                    "audio"
                ],
                "summary": "RNNoise/afftdn denoise -> 48 kHz mono WAV",
                "operationId": "denoise_v1_audio_denoise_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DenoiseIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AudioOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/jobs/{job_id}": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Poll a render job",
                "description": "`status` is queued | running | done | error.\n\nOn `done` the payload is under `result`. On `error` it is under `error`.\nJobs are kept for JOB_TTL_SECONDS after they finish, then 404.",
                "operationId": "get_job_v1_jobs__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobStatus"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AudioOut": {
                "properties": {
                    "audio": {
                        "type": "string",
                        "title": "Audio",
                        "description": "base64 WAV, 48 kHz mono"
                    }
                },
                "type": "object",
                "required": [
                    "audio"
                ],
                "title": "AudioOut"
            },
...
            "DenoiseIn": {
                "properties": {
                    "audio": {
                        "type": "string",
                        "title": "Audio"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "audio"
                ],
                "title": "DenoiseIn"
            },
...
            "EditIn": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "op": {
                        "$ref": "#/components/schemas/EditOp"
                    },
                    "mask": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mask",
                        "description": "base64/URL PNG. White = the region to regenerate."
                    },
                    "params": {
                        "$ref": "#/components/schemas/EditParams"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "image",
                    "op"
                ],
                "title": "EditIn"
            },
            "EditOp": {
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "restore",
                            "colorize",
                            "replace-sky",
                            "magic-eraser",
                            "generative-fill",
                            "prompt",
                            "upscale",
                            "remove-background"
                        ],
                        "title": "Type"
                    },
                    "prompt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1000
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prompt"
                    },
                    "factor": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "enum": [
                                    2,
                                    4
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Factor"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "type"
                ],
                "title": "EditOp"
            },
            "EditParams": {
                "properties": {
                    "negative_prompt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 300
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Negative Prompt"
                    },
                    "strength": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 1.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Strength"
                    },
                    "steps": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 60.0,
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Steps"
                    },
                    "guidance_scale": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 20.0,
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guidance Scale"
                    },
                    "seed": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 2147483647.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seed"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "title": "EditParams",
                "description": "Optional diffusion tuning. Every field is clamped, never passed through raw.\n\nSame bounds as the old `sanitizeParams` in edit/route.ts \u2014 a caller cannot\nask for 500 steps and pin the GPU."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ImageOut": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image",
                        "description": "base64 PNG"
                    },
                    "width": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Width"
                    },
                    "height": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Height"
                    }
                },
                "type": "object",
                "required": [
                    "image"
                ],
                "title": "ImageOut"
            },
            "JobAccepted": {
                "properties": {
                    "job_id": {
                        "type": "string",
                        "title": "Job Id"
                    },
                    "status": {
                        "type": "string",
                        "const": "queued",
                        "title": "Status",
                        "default": "queued"
                    },
                    "poll": {
                        "type": "string",
                        "title": "Poll"
                    }
                },
                "type": "object",
                "required": [
                    "job_id",
                    "poll"
                ],
                "title": "JobAccepted"
            },
            "JobStatus": {
                "properties": {
                    "job_id": {
                        "type": "string",
                        "title": "Job Id"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "queued",
                            "running",
                            "done",
                            "error"
                        ],
                        "title": "Status"
                    },
                    "created_at": {
                        "type": "number",
                        "title": "Created At"
                    },
                    "started_at": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Started At"
                    },
                    "finished_at": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Finished At"
                    },
                    "result": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Result"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    }
                },
                "type": "object",
                "required": [
                    "job_id",
                    "status",
                    "created_at"
                ],
                "title": "JobStatus"
            },
...
            "OcrIn": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "image"
                ],
                "title": "OcrIn"
            },
            "OcrLine": {
                "properties": {
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "score": {
                        "type": "number",
                        "title": "Score"
                    },
                    "box": {
                        "items": {
                            "items": {
                                "type": "number"
                            },
                            "type": "array"
                        },
                        "type": "array",
                        "title": "Box"
                    }
                },
                "type": "object",
                "required": [
                    "text",
                    "score",
                    "box"
                ],
                "title": "OcrLine"
            },
            "OcrOut": {
                "properties": {
                    "lines": {
                        "items": {
                            "$ref": "#/components/schemas/OcrLine"
                        },
                        "type": "array",
                        "title": "Lines"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    }
                },
                "type": "object",
                "required": [
                    "lines",
                    "text"
                ],
                "title": "OcrOut"
            },
            "RemoveBgIn": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "model_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Name"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "image"
                ],
                "title": "RemoveBgIn"
            },
            "Segment": {
                "properties": {
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "start_sec": {
                        "type": "number",
                        "title": "Start Sec"
                    },
                    "end_sec": {
                        "type": "number",
                        "title": "End Sec"
                    }
                },
                "type": "object",
                "required": [
                    "text",
                    "start_sec",
                    "end_sec"
                ],
                "title": "Segment"
            },
            "TranscribeIn": {
                "properties": {
                    "audio": {
                        "type": "string",
                        "title": "Audio"
                    },
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "timestamps": {
                        "type": "boolean",
                        "title": "Timestamps",
                        "default": false
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "audio"
                ],
                "title": "TranscribeIn"
            },
            "TranscribeOut": {
                "properties": {
                    "transcript": {
                        "type": "string",
                        "title": "Transcript"
                    },
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "segments": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Segment"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Segments"
                    }
                },
                "type": "object",
                "required": [
                    "transcript"
                ],
                "title": "TranscribeOut"
            },
...
            "UpscaleIn": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "scale": {
                        "type": "integer",
                        "enum": [
                            2,
                            4
                        ],
                        "title": "Scale",
                        "default": 4
                    },
                    "face_enhance": {
                        "type": "boolean",
                        "title": "Face Enhance",
                        "default": false
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "image"
                ],
                "title": "UpscaleIn"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}